chore: extrait la fonction créerUtilisateursAvecDroits (000)
