+++ src/database/seed.ts
@@ -10,7 +10,6 @@ import {
   synthese_des_resultats,
 } from '@prisma/client';
 import { faker } from '@faker-js/faker/locale/fr';
-import { v4 as uuidv4 } from 'uuid';
 
 import SynthèseDesRésultatsSQLRowBuilder
   from '@/server/infrastructure/test/builders/sqlRow/SynthèseDesRésultatsSQLRow.builder';
@@ -25,6 +24,11 @@ import CommentaireRowBuilder from '@/server/infrastructure/test/builders/sqlRow/
 import IndicateurRowBuilder from '@/server/infrastructure/test/builders/sqlRow/IndicateurSQLRow.builder';
 import ObjectifSQLRowBuilder from '@/server/infrastructure/test/builders/sqlRow/ObjectifSQLRow.builder';
 import { DIR_CHANTIER, DIR_PROJET, DITP_ADMIN, DITP_PILOTAGE, PM } from '@/server/domain/identité/Profil';
+import {
+  créerUtilisateursAvecDroits,
+  InputProfil,
+  InputUtilisateur,
+} from '@/server/infrastructure/accès_données/identité/seed';
 
 const prisma = new PrismaClient();
 
@@ -153,7 +157,7 @@ class DatabaseSeeder {
   }
 
   private async _créerDroits() {
-    const inputProfils = [
+    const inputProfils: InputProfil[] = [
       { code: PM, nom: 'Premier Ministre', a_acces_tous_chantiers: true },
       { code: DITP_ADMIN, nom: 'DITP - Admin', a_acces_tous_chantiers: true },
       { code: DITP_PILOTAGE, nom: 'DITP - Pilotage', a_acces_tous_chantiers: true },
@@ -164,7 +168,7 @@ class DatabaseSeeder {
     // noinspection TypeScriptValidateTypes
     const chantiersRows = await prisma.chantier.findMany({ distinct: ['id'], select: { id: true }, take: 10 });
     const chantierIds = chantiersRows.map(it => it.id);
-    const inputUtilisateurs = [
+    const inputUtilisateurs: InputUtilisateur[] = [
       { email: 'utilisateur_PM@example.com', profilCode: PM, chantierIds: [] },
       { email: 'utilisateur_DITP_A@example.com', profilCode: DITP_ADMIN, chantierIds: [] },
       { email: 'utilisateur_DITP_P@example.com', profilCode: DITP_PILOTAGE, chantierIds: [] },
@@ -172,67 +176,7 @@ class DatabaseSeeder {
       { email: 'utilisateur_DIR_P@example.com', profilCode: DIR_PROJET, chantierIds },
     ];
 
-    type DonnéesProfil = { id: string, code: string, nom: string, a_acces_tous_chantiers: boolean };
-    const donnéesProfils: Record<string, DonnéesProfil> = {};
-    for (const input of inputProfils) {
-      donnéesProfils[input.code] = { ...input, id: uuidv4() };
-    }
-
-    type DonnéesUtilisateur = { id: string, email: string, profil_id: string, chantier_ids: string[] };
-    const donnéesUtilisateurs: DonnéesUtilisateur[] = [];
-    for (const input of inputUtilisateurs) {
-      donnéesUtilisateurs.push({
-        id: uuidv4(),
-        email: input.email,
-        profil_id: donnéesProfils[input.profilCode].id,
-        chantier_ids: input.chantierIds,
-      });
-    }
-
-    for (const profil of Object.values(donnéesProfils)) {
-      await prisma.profil.create({ data: profil });
-    }
-
-    for (const { id, email, profil_id } of Object.values(donnéesUtilisateurs)) {
-      await prisma.utilisateur.create({ data: { id, email, profil_id } });
-    }
-
-    const utilisateurChantiers = [];
-    for (const { id: utilisateur_id, chantier_ids } of donnéesUtilisateurs) {
-      for (const chantier_id of chantier_ids) {
-        utilisateurChantiers.push({
-          utilisateur_id,
-          chantier_id,
-        });
-      }
-    }
-    await prisma.utilisateur_chantier.createMany({ data: utilisateurChantiers });
-
-    const { id: habilitationScopeLectureId } = await prisma.habilitation_scope.create({
-      data: {
-        code: 'lecture',
-        nom: 'Scope de lecture sur un chantier',
-      },
-    });
-
-    const { id: habilitationScopeÉcritureId } = await prisma.habilitation_scope.create({
-      data: {
-        code: 'écriture',
-        nom: "Scope d'écriture sur un chantier",
-      },
-    });
-
-    const profilHabilitations = [
-      { profil_id: donnéesProfils.PM.id, habilitation_scope_id: habilitationScopeLectureId },
-    ];
-    for (const code of [DITP_ADMIN, DIR_CHANTIER]) {
-      for (const habilitation_scope_id of [habilitationScopeLectureId, habilitationScopeÉcritureId]) {
-        const profil_id = donnéesProfils[code].id;
-        profilHabilitations.push({ profil_id, habilitation_scope_id });
-      }
-    }
-
-    await prisma.profil_habilitation.createMany({ data: profilHabilitations });
+    await créerUtilisateursAvecDroits(prisma, inputProfils, inputUtilisateurs);
   }
 }
